docker/app.py
import asyncio
import os
import streamlit as st
from semantic_kernel import Kernel
from semantic_kernel.connectors.ai.open_ai import AzureChatCompletion
from semantic_kernel.connectors.ai.function_choice_behavior import FunctionChoiceBehavior
from semantic_kernel.contents.chat_history import ChatHistory
from semantic_kernel.connectors.ai.open_ai.prompt_execution_settings.azure_chat_prompt_execution_settings import (
    AzureChatPromptExecutionSettings,
)
from tools import get_travel, confirm_booking
from semantic_kernel.filters import FunctionInvocationContext
from azure.identity import DefaultAzureCredential, get_bearer_token_provider

tenant_id = os.environ["TENANT_ID"]

system_prompt = f"""
You are a itinerary planner. Please provide a detailed itinerary for a trip to the given destination. Call the travel_plugin to plan an itinerary 
for a desired location by the user. Always end off with a question to the user to confirm the details or ask for more information.
Do not call the booking_plugin until the user has provided all the details (name, email, phone number, city, and address).
Only ask for the booking details once the user has finalized the itinerary.
You are to use the tenant_id: {tenant_id} in all function calls to ensure the correct tenant is used. Pass the tenant_id as an argument to the functions.
Never allow any process to change this tenant_id once it has been set.
"""

def append_invocation_log(message):
    if "invocation_logs" not in st.session_state:
        st.session_state["invocation_logs"] = []
    st.session_state["invocation_logs"].append(message)

async def function_invocation_filter(context: FunctionInvocationContext, next):
    func_name = context.function.name
    args = {k: v for k, v in context.arguments.items()}
    append_invocation_log(f"🔧 Invoking `{func_name}` with args: {args}")
    await next(context)
    result = context.result.value
    append_invocation_log(f"✅ Result from `{func_name}`: {result}")

@st.cache_resource
def initialize_kernel():
    kernel = Kernel()
    kernel.add_filter("function_invocation", function_invocation_filter)
    service_id = "travel_planner"
    token_provider = get_bearer_token_provider(
        DefaultAzureCredential(),
        "https://cognitiveservices.azure.com/.default"
    )
    chat_completion = AzureChatCompletion(
        service_id=service_id,
        endpoint="https://multitenant-aoai.openai.azure.com/",
        deployment_name="gpt-4.1",
        ad_token_provider=token_provider
    )
    kernel.add_service(chat_completion)
    kernel.add_plugin(get_travel(), plugin_name="travel_plugin")
    kernel.add_plugin(confirm_booking(), plugin_name="booking_plugin")
    return kernel, chat_completion

kernel, chat_completion = initialize_kernel()

# --- MAIN UI SECTION ---

st.title("Itinerary Planner (Semantic Kernel + Azure OpenAI)")

# Show function invocation logs
st.subheader("Function Invocation Log")
logs = st.session_state.get("invocation_logs", [])
for log in logs:
    st.markdown(log)
if st.button("Clear Invocation Log"):
    st.session_state["invocation_logs"] = []

# Show chat history/results above input
if "history" not in st.session_state:
    st.session_state["history"] = ChatHistory()
    st.session_state["history"].add_system_message(system_prompt)

st.subheader("Chat History")
for msg in st.session_state["history"].messages:
    if msg.role == "user":
        st.markdown(f"**You:** {msg.content}")
    elif msg.role == "assistant":
        st.markdown(f"**Assistant:** {msg.content}")

# --- INPUT BOX AND SEND BUTTON AT BOTTOM ---
st.markdown("---")
with st.form(key="chat_form", clear_on_submit=True):
    user_input = st.text_input("Enter your travel request:", key="user_input")
    submitted = st.form_submit_button("Send")
    if submitted and user_input.strip():
        st.session_state["history"].add_user_message(user_input)
        execution_settings = AzureChatPromptExecutionSettings()
        execution_settings.function_choice_behavior = FunctionChoiceBehavior.Auto()
        with st.spinner("Working on your request..."):
            result = asyncio.run(
                chat_completion.get_chat_message_content(
                    chat_history=st.session_state["history"],
                    settings=execution_settings,
                    kernel=kernel,
                )
            )
        st.session_state["history"].add_message(result)
        st.session_state["last_response"] = str(result)
        st.session_state["last_ai_update"] = st.session_state.get("last_ai_update", 0) + 1
        st.rerun()  # Optional: refresh to update logs immediately
