docker/requirements.txt
azure-identity
azure-storage-blob
semantic-kernel
azure-ai-projects
streamlit